RowBox fix and BoxExpression refactor...
Move BoxExpression out of mathics.builtin.base and into
mathics.box.expression

BoxExpression didn't have has_form and therefore RowBoxing was giving us
too-small boxes when put in a Table.

BoxExpression now has .elements just like an Expression does.

Deprecated BoxConstruct removed.
