+++ mathics/builtin/base.py
@@ -10,26 +10,18 @@ from typing import Any, Callable, Dict, Iterable, List, Optional, Union, cast
 import sympy
 
 from mathics.core.atoms import Integer, MachineReal, PrecisionReal, String
-from mathics.core.attributes import A_NO_ATTRIBUTES, A_PROTECTED, A_READ_PROTECTED
+from mathics.core.attributes import A_NO_ATTRIBUTES, A_PROTECTED
 from mathics.core.convert.expression import to_expression, to_numeric_sympy_args
 from mathics.core.convert.op import ascii_operator_to_symbol
 from mathics.core.convert.python import from_bool
 from mathics.core.convert.sympy import from_sympy
 from mathics.core.definitions import Definition
-from mathics.core.element import BoxElementMixin
 from mathics.core.exceptions import MessageException
 from mathics.core.expression import Expression, SymbolDefault
-from mathics.core.list import ListExpression
 from mathics.core.number import PrecisionValueError, get_precision
 from mathics.core.parser.util import PyMathicsDefinitions, SystemDefinitions
 from mathics.core.rules import BuiltinRule, Pattern, Rule
-from mathics.core.symbols import (
-    BaseElement,
-    Symbol,
-    SymbolHoldForm,
-    ensure_context,
-    strip_context,
-)
+from mathics.core.symbols import BaseElement, Symbol, ensure_context, strip_context
 from mathics.core.systemsymbols import SymbolMessageName, SymbolRule
 
 # Signals to Mathics doc processing not to include this module in its documentation.
@@ -73,28 +65,6 @@ def get_option(options, name, evaluation, pop=False, evaluate=True):
 
 def has_option(options, name, evaluation):
     return get_option(options, name, evaluation, evaluate=False) is not None
-
-
-def split_name(name: str) -> str:
-    """
-    insert spaces in front of upper case letters
-    and numbers. For instance,
-    ``split_name("BezierCurve3D")`` results in
-    ``"bezier curve 3D"``
-
-    """
-    if name == "":
-        return ""
-    result = name[0]
-    for i in range(1, len(name)):
-        if name[i].isupper():
-            if not name[i - 1].isdigit():
-                result = result + " "
-        elif name[i].isdigit():
-            if not name[i - 1].isdigit():
-                result = result + " "
-        result = result + name[i]
-    return result.lower()
 
 
 mathics_to_python = {}  # here we have: name -> string
@@ -718,162 +688,6 @@ class SympyFunction(SympyObject):
         return sympy_expr
 
 
-class BoxExpression(BuiltinElement, BoxElementMixin):
-    # This is the base class for the "Final form"
-    # of formatted expressions.
-    #
-    # These objects should not be evaluated, so in the evaluation process should be
-    # considered "inert". However, it could happend that an Expression having them as an element
-    # be evaluable, and try to apply rules. For example,
-    # InputForm[ToBoxes[a+b]]
-    # should be evaluated to ``Expression(SymbolRowBox, '"a"', '"+"', '"b"')``.
-    #
-    # Changes to do, after the refactor of mathics.core:
-    #
-
-    # * Review the implementation
-
-    attributes = A_PROTECTED | A_READ_PROTECTED
-
-    def __new__(cls, *elements, **kwargs):
-        instance = super().__new__(cls, *elements, **kwargs)
-        # This should not be here.
-        article = (
-            "an "
-            if instance.get_name()[0].lower() in ("a", "e", "i", "o", "u")
-            else "a "
-        )
-
-        instance.summary_text = (
-            "box representation for "
-            + article
-            + split_name(cls.get_name(short=True)[:-3])
-        )
-        if not instance.__doc__:
-            instance.__doc__ = rf"""
-            <dl>
-              <dt>'{instance.get_name()}'
-              <dd> box structure.
-            </dl>
-            """
-
-        # the __new__ method from BuiltinElement
-        # calls self.init. It is expected that it set
-        # self._elements. However, if it didn't happens,
-        # we set it with a default value.
-        # There should be a better way to implement this
-        # behaviour...
-        if not hasattr(instance, "_elements"):
-            instance._elements = tuple(elements)
-        return instance
-
-    @property
-    def is_literal(self) -> bool:
-        """
-        True if the value can't change, i.e. a value is set and it does not
-        depend on definition bindings. That is why, in contrast to
-        `is_uncertain_final_definitions()` we don't need a `definitions`
-        parameter.
-        """
-        return False
-
-    @property
-    def elements(self):
-        return self._elements
-
-    @elements.setter
-    def elements(self, value):
-        self._elements = value
-        return self._elements
-
-    def tex_block(self, tex, only_subsup=False):
-        if len(tex) == 1:
-            return tex
-        else:
-            if not only_subsup or "_" in tex or "^" in tex:
-                return "{%s}" % tex
-            else:
-                return tex
-
-    def to_expression(self) -> Expression:
-        # FIXME: All classes should store their symbol name.
-        # So there should be a self.head.
-        return Expression(Symbol(self.get_name()), *self._elements)
-
-    def replace_vars(self, vars, options=None, in_scoping=True, in_function=True):
-        expr = self.to_expression()
-        result = expr.replace_vars(vars, options, in_scoping, in_function)
-        return result
-
-    # Deprecated: remove eventually
-    def get_elements(self):
-        return self._elements
-
-    def get_head_name(self):
-        return self.get_name()
-
-    def get_lookup_name(self):
-        return self.get_name()
-
-    def get_sort_key(self) -> tuple:
-        return self.to_expression().get_sort_key()
-
-    def get_string_value(self):
-        return "-@" + self.get_head_name() + "@-"
-
-    def sameQ(self, expr) -> bool:
-        """Mathics SameQ"""
-        return expr.sameQ(self)
-
-    def do_format(self, evaluation, format):
-        return self
-
-    def format(self, evaluation, fmt):
-        expr = Expression(SymbolHoldForm, self.to_expression())
-        fexpr = expr.format(evaluation, fmt)
-        return fexpr
-
-    def get_head(self):
-        return Symbol(self.get_name())
-
-    @property
-    def head(self):
-        return self.get_head()
-
-    @head.setter
-    def head(self, value):
-        raise ValueError("BoxExpression.head is write protected.")
-
-    @property
-    def leaves(self):
-        return self.get_elements()
-
-    @leaves.setter
-    def leaves(self, value):
-        raise ValueError("BoxExpression.elements is write protected.")
-
-    def flatten_pattern_sequence(self, evaluation) -> "BoxExpression":
-        return self
-
-    def flatten_with_respect_to_head(self, symbol):
-        return self.to_expression().flatten_with_respect_to_head(symbol)
-
-    def get_option_values(self, elements, **options):
-        evaluation = options.get("evaluation", None)
-        if evaluation:
-            default = evaluation.definitions.get_options(self.get_name()).copy()
-            options = ListExpression(*elements).get_option_values(evaluation)
-            default.update(options)
-        else:
-            from mathics.core.parser import parse_builtin_rule
-
-            default = {}
-            for option, value in self.options.items():
-                option = ensure_context(option)
-                default[option] = parse_builtin_rule(value)
-        return default
-
-
 class PatternError(Exception):
     def __init__(self, name, tag, *args):
         super().__init__()
@@ -1034,7 +848,3 @@ class CountableInteger:
                             return CountableInteger(0, upper_limit=True)
 
         return None  # leave original expression unevaluated
-
-
-# Backward compatibility
-BoxConstruct = BoxExpression
